PID controller tuning (#8)
* Count frequency change frame as 0th frame

* Increase PID controller's bootstrap phase to 100 frames

* Simplify PID controller interface

* Fix whitespace

* Control number of frames to skip instead of frequency in PID controller

* Update PID constants

* Remove redundant PIDController function

* Throw runtime error instead of abort in case of illegal frequency mode

+++ InfiniTAM/Apps/InfiniTAM_cli/CLIEngine.h
@@ -35,7 +35,7 @@ namespace InfiniTAM
 				unsigned freqDivisor;
 				int framesSinceFreqChange;
 				std::vector<unsigned> *processed;
-				std::vector<unsigned> *frequencies;
+				std::vector<double> *frequencies;
 				std::vector<unsigned> *newBricks;
 			};
